pyproject.toml
[project]
name = "cyber-bible"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "eel>=0.18.2",
    "instagrapi>=2.3.0",
    "pillow>=12.1.1",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
]
